+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 * Cleaned up Github Actions workflows.
 * Updated PHP dependencies.
 * Added Playwright github action.
+* Changed how templates are imported.
 
 ### NB! Prior to 3.x the project was split into separate repositories